backend/lambda/tableau/requirements.txt
boto3==1.18.31
botocore==1.21.31
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
Deprecated==1.2.13
idna==3.2
jmespath==0.10.0
numpy==1.22.1
pandas==1.3.5
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sql-metadata==2.2.2
sqlparse==0.4.1
tableauserverclient==0.16.0
urllib3==1.26.6
wrapt==1.12.1
